extract SearchBar
