+++ frontend/src/metabase/nav/constants.js
@@ -1,0 +1,2 @@
+import { color, lighten } from "metabase/lib/colors";
+export const DefaultSearchColor = lighten(color("nav"), 0.07);
